fileserver/requirements.txt
flask==3.0.0
pandas==2.1.3
openpyxl==3.1.2
waitress==2.1.2
python-dotenv==1.0.0